Add clip LED to "progress bar" level meter
Create a new class to gather the widgets of the progress bar meter (two progress
bars: the same old one, and one more for the clipping indication).

The fancy skin LED display could also be grouped in a similar class.
CMultiColorLEDBar would then just become an interface to either that class or
the one created in this commit.