debian: Trim build dependencies
dpkg-dev is not needed for building binaries.

make, g++ and libc6-dev will be pulled in by e.g. build-essential
which is not necessary to specify (Debian policy).

Packages should not build-depend on qt5-default but rather
on qt5-base-dev and select QT5 in debian/rules
(https://wiki.debian.org/Lintian/Tags/depends-on-metapackage)

qtdeclarative5-dev is simply not needed.

qttools5-dev-tools is needed for translations using lrelease.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>