Implement new RX jitter buffer for use with Jamulus.
The new jitter buffer uses a timing histogram to compute the actual
jitter relative to the audio device or system clock timer.
This will give a more accurate jitter value.

A sequence bit has been added to all transmitted frames to improve
the clock drift computation in the analyzer console.

Signed-off-by: Hans Petter Selasky <hps@selasky.org>