revert removal of binary *.qm files from .tar.gz distribution:
as discussed in https://github.com/corrados/jamulus/pull/799
these files need to be present in the .tar.gz distribution until
we can require every user to either have a working lrelease.prf
(Qt 5.12 and above, but seemingly not working right under Windows®
either) or to manually run lrelease before building, as is done
e.g. for the Debian backport to buster:

diff --git a/debian/rules b/debian/rules
index 7d93ed2..5234e86 100755
--- a/debian/rules
+++ b/debian/rules
@@ -45,4 +49,5 @@ endif
 	dh $@
 
 override_dh_auto_configure:
+	cd src/res/translation && lrelease *.ts
 	dh_auto_configure -- ${CONFIG_ARGS}

----

This is a separate commit from its parent so, when the time
comes, it’ll be easy to “git revert” it.