qmake: avoid unnecessary runtime dependencies
Most g++ versions already seem to use --as-needed in the ld invocations
by default. Some versions such as Debian buster's don't. This leads to
an unnecessary runtime dependency on libQt5Concurrent.so.

This PR ensures that qmake passes --as-needed in all cases to avoid
this.

Fixes #741.
Verified build on: Arch Linux, g++ 10.2.0
Verified build on: Debian buster, g++ 8.3.0

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>