More generic structure for responding to MIDI controller messages
This is preparation for continuing work on issue #95.  While it does
not yet introduce functional differences, MIDI controller information
is organised in a manner where reacting to more than just fader
control messages becomes easy.

The preparation right now is for controllers of type fader, pan, solo,
mute but can be easily extended for other messages arriving on a
single channel.