Implement Solo and Mute via MIDI Controller
This implements the --ctrlmidich controller specifications 's' and
'm'.  Since the MIDI connection is onedirectional and it is not clear
how controller state may be maintained for headless operation, this
currently relies on the controller maintaining and displaying
state (and accepting that changing the state via the GUI will let the
states of controller and GUI diverge).

It's not clear to me what this implies for headless operation and what
kind of code might need to get added to make that work.