Server: Stop using Cloudflare for IP detection
Previously, the code for finding the current machine's external
IP address created a TCP connection to Cloudflare. This is unnecessary.
Instead, just creating a socket to a non-private IP suffices to find
that IP. We now use a reserved IP for that and use UDP in order to avoid
triggering any traffic at all.

All credit goes to @atsampson.
Fixes #633.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>