Fix headless build
The qmake definition incorrectly listed src/multicolorled.h in the
global HEADERS list. However, that file includes QLabel which is only
available when building with QT += widgets. This is not done in headless
builds. Therefore, the build breaks.

This commit moves the include to the HEADERS_GUI list and fixes the
build.

src/multicolorled.h is only included by UI-related classes. It should
therefore be safe to move it.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>