Codacy/Flawfinder: Document a false positive finding
The read() from the signal handler socket triggers a Codacy/flawfinder finding:
"Check buffer boundaries if used in a loop including recursive loops (CWE-120, CWE-20)"
Investigation has shown that this is a false positive as the proper buffer
size is enforced and because the input is within the same security boundary.

This adds an explaining comment and a special hint at Codacy/Flawfinder to ignore
the (non-)problematic line.

Related: #1081

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>