Server: Use a traffic-less way to get external IP
Previously, the code for finding the current machine's external
IP address created a TCP connection to Cloudflare's DNS. This is unnecessary.
Instead, just creating a socket to a non-private IP suffices to find
that IP. We still use Cloudflare's IP, but we will no longer send any traffic
there as we use an UDP socket without actually sending any data.

Credit for finding this approach goes to @atsampson.
Fixes #633.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>