Actually use the machine specific OPUS code
Set defines so that the preprocessor doesn't drop SSE code, add some
compiler flags (for gcc) to enable SSE instructions.  Note that Opus
already has code to detect the support for SSE available at runtime,
so that the SEEx code compiled in the specialized files will be only
called when it can actually work.