Revert "Client: Always update gain levels"
This reverts PR #1009 (f116811a4e9748722a2f1b8178ebfe68737b57d0) as it
has unintended side effects (#1191).

Basically, that PR assumed that applying stored fader levels would
always be a safe thing to do.
This is not the case, as changes in fader levels are currently not
stored immediately, but rather in specific events only (e.g.
disconnects).

PR #1009 was initially supposed to help with #955 as a client-side fix.
With the revert, we still have the server-side fix from #1010.

It is planned to re-apply #1009 once fader-level storage has been fully
understood and extended to cover all relevant places (#1191).