Generate .qm files during build
This removes .qm files from git, adds them to .gitignore and changes
Jamulus.pro to generate the files dynamically during build in a way
which hopefully even works on ancient Qt versions.

This makes use of custom compiler support in qmake which seems to be
available (at least) as of Qt 5.5.
https://doc.qt.io/archives/qt-5.5/qmake-advanced-usage.html#adding-compilers

Fixes #1111