Windows: Delay detaching from console
This moves the call to attach to the parent process' console from the
start of the main routine to a place after command line parsing.

The goal is to improve command-line-only invocations of Jamulus for
non-headless builds, e.g. `Jamulus -h` or `Jamulus -n -s`.

It also modifies the logic to disable the detaching not only for
headless builds but also for non-headless builds with --nogui
invocations.

Fixes #1281.