Favorites update
Keeping original structure so that private servers and port forwarded
servers will work without any Directory server intervention.  Thus they
can be used when the Directory servers are striking (which continues to
happen, not alot, but all too often).

If no ping is received from a server a server list request is made to
it's Directory server (max. 1 request per Directory server).  Then pings
are requested again.
When a server list is seen from a Directory server the stored IP and max
users are checked and updated in the Favorites if necessary.
Requests are send 1 per OnTimerPing so as to keep down network traffic.
Servers are marked "offline" if after a server list request and
following ping requests there is still no ping received.
The entire sequence can be restarted by reopening the Favorites tab or
reopening the window.

The sort order is last used or added server at the top of the list
pushing down the previous servers.

Unsolved problem: if there are 2 servers with the same name on the same
Directory server the IP update may update the wrong one.