Server: Sync channel list right after connect
Previously, the list of active channels was only distributed to clients
when metadata such as names had changed. This allowed for clients to remain
in a semi-connected state.

This change ensures that a complete channel list is sent as soon as a
new client connects. Therefore, state is properly synced across all
clients.

Fixes #1243.