Do not hardcode fontheight
Previously we hardcoded a fixed label height to prevent flickering when
muting someone. This seemed to work on Windows, but the value was wrong
for certain fonts on Linux and other operating systems.

Now we use the existing font height of the "Pan" label in order to use a
safe fontsize.

Fixes #1312
Fixes #811 for real now