Add button to reset device
This lets Jamulus become aware of new configurations after the device
is changed from a non-working to working state.  Technically, this was
already by possible by reselecting the device in the combobox.  But
that is a confusing and very non-obvious interface.

Also, if the settings don't have a saved selected device, try to
autoselect one.