Measurements placed on clientdlg
The values are not updated on the cleintsettings tab.
I don't think this is too ugly, so it could be workable.
Please test to see if it no longer has memory loss.
The only problem I see is with the stream data rate which is now not
shown.  I could imagine to use a static variable to store the last value
and only update when changed, which will be almost never.

I will start now to clean this up with the accessibility stuff, etc.
where it should be.  Will not change the functioning.