Mac autobuild: Update to Xcode 12
Let the autobuild default to Xcode 12 again, but still ensure
that we limit SDK support to 10.15 due to lack of Qt-wise support
for newer versions.

This is a slightly different approach to the already merged #1655.

We are limited to <= Xcode-12.1.1 though because we require SDK 10.15
which is not shipped on more recent versions anymore.
This change has a (low) chance of fixing recent Mac issues on Big Sur
(#1643, #1791). At least, it makes us use the latest officially
recommended setup.

This only touches the regular Mac build, not the legacy build.

Support matrix:
https://doc.qt.io/qt-5/macos.html

Variables:
https://doc.qt.io/qt-5/qmake-variable-reference.html#qmake-mac-sdk

Explanations:
https://doc.qt.io/qt-5/macos-deployment.html