Multichannel Input using phantom clients
Dual Mono-in/Stereo out added.  Note this is not yet working for delay
panned servers.  I don't understand the code, still looking on that.

There is no change to the protocol, works with old clients.
Using version number 3.9.9 and requires that on the server to enable at
the client.

Also compiled with the new ASIOSDK2.3.3 seems to be working fine.

I'm thinking an addition to the protocol between the DualMono input
client and server might be nice to be able to give the phantom channel a
different name and instrument. That could still work with old clients at
the receiving end.  Right now I am just appending -L and -R to the name.