Remove unnecessary submodules from tools directory
Git submodules are intended for bringing in a copy of another repo
that is needed for building the project. They are not for general
reference to other projects of interest.

This commit removes those unneeded submodules, and instead adds a text
file listing those repos, with a link to each and a brief description.