tools: Add changelog-helper.sh
The helper verifies that the ChangeLog file contains references to
all PRs from `git log` and to all PRs from the relevant Github
milestone.

It can also automatically generate entries for PRs.

It respects the magic `CHANGELOG:` keyword in PRs and PR comments to
allow users to help fill this.
It respects the magic `CHANGELOG: SKIP` term to avoid a ChangeLog entry
for a given PR.

It supports grouping and sorting within the groups.

Relevant discussion: #1839

Co-authored-by: ann0see <20726856+ann0see@users.noreply.github.com>