Add tools/update-copyright-notices and run via CI
This script updates all global copyright headers to the current year (2022).
It also updates all source file's copyright headers according to year of
the most recent commit.

It is set up to run on all pushes to master (including PR merges).
If there's something to do, a PR is opened from a temporary branch.
If the PR is closed, the temporary branch will be deleted automatically.

Co-authored-by: Tony Mountifield <tony@mountifield.org>