Mac: Use native Dock Badge handling
This commit replaces the deprecated use of Qt5 macextras by a small ObjC
integration into the native methods.
The macextras build option and dependency is removed accordingly.

This is required for Qt6 compatibility, because Qt6 no longer ships
macextras and suggests going the path that this commit goes.

To avoid duplicated maintenance, the switch is done indepdendent of the
Qt version as the new code should work on Qt5 as well.