Audiomixerboard: Adjust size when switching meter styles
There are two different base meter styles (LED vs. bars) with multiple
variants each (regular, slim, round).
The base meter styles are implmeneted as widgets and are always part of
a QStackedLayout. The selected meter style is shown via
setCurrentWidget(). This hides the inactive widget, but the inactive
widget is still there and is part of QStackedLayout's size calculations
which always uses the maximum size over all widgets, not just the
currently visible one.

This leads to problems when the currently visible meter widget is
narrower than the hidden counterpart. In that case, the QStackedLayout
is wider than expected.

Technically, the following transitions show this issue:
- Select Bar, then Small LEDs, or
- Select LEDs, then Narrow Bar

However, the latter case already had a workaround (choosing Narrow Bar
silently adjusted details of the LED meter as well).

The proper fix is using a subclass of QStackedLayout which calculates its
sizeHint solely based on the currently visible widget.
This fixes the Bar->Small LEDs transition and makes the LEDs->Narrow Bar
workaround redundant, which has therefore been removed.

Fixes #2351