Autobuild: Move ios build scripts up one level
There's no need to place the ios autobuild scripts in an "artifacts" sub
directory. This was probably done as the ios build was likely based on
the mac build logic.

This commit moves the .sh files to their new location and adjusts a single
line to account for the reduced nesting level. It also updates the callers
to use the new paths.