Coding style: Require clang-format 13
We currently use clang-format 10 which has shown to produce unexpected
output. This is annoying for PR submitters as the CI forces non-optimal
code layout, especially when submitters use a later clang-format version
which does not show this behavior.

This commit changes CI to use clang-format 13 and updates relevant
clang-format inline docs.

This also updates the version of the Github Action to the latest version
as this is required for v13 support.

Related: #2552