Autobuild: Move Changelog generation to the workflow
Previously, the autobuild changelog (autoLatestChangelog.md) which is
used for the Github release body, was generated as part of the
"analyse_git_reference.py" script. It does not seem logical for an
analysis script to have a side effect of producing a Changelog file.
Also, the file name (autoLatestChangelog.md) was hardcoded in two places
and had to match (workflow + script).

With this commit, the changelog generation call (.github/actions_scripts/getChangelog.pl)
is moved to the workflow. This follows the principle of least surprise
and moves the output file name usage to a single file (the workflow).

Fixes: #2480