Coding style: Add .editorconfig
This is a new config file which is supported by multiple editors and
IDEs in order to get started with a consistent style, e.g. regarding
indentations.
This makes sense for both file types which are covered by linters (.cpp,
.sh) as it provides a sane starting point and has additional features
(e.g. config for line endings) as well as for file types which do not
have linting at all.
Having such a generic config file seems preferable to having none or
multiple editor-specific configs.