Autobuild: Limit build targets by default
This modifies the build logic to stop building all platforms on PRs by
default. To re-enable building for all platforms, the PR has to contain
a special hint which has been documented in the PR template.
Manually triggered workflows can also decide this explicitly now.

Tags/releases and autobuild branches are unchanged.

This is supposed to reduce CI usage times in order to keep build queues
for us shorter and in order to avoid wasting energy.