Build: Force cross-platform main() usage on MSVC with Qt6
Qt5 had a special qtmain library which took care of forwarding
the MSVC default WinMain() entrypoint to the platform-agnostic main().
Qt6 is still supposed to have that lib under the new name QtEntryPoint.
As it does not seem to be effective when building with qmake, we are rather
instructing MSVC to use the platform-agnostic main() entrypoint directly.

Without this change, building fails.