CI: Switch Linux Build to Ubuntu 22.04 runner
We continue using an Ubuntu 18.04 image via containers.
This requires slight adjustments in the Matrix logic and an
additional step to install some dependencies (sudo, git) which are
usually pre-installed on Github's images.

It also causes LD_PRELOAD ERROR messages due to CodeQL.
Sadly, there does not seem a way to avoid that.
Based on the logs, CodeQL still works properly.
The most likely reason for the errors is that CodeQL is installed within
the container, but the environment variables for LD_PRELOAD have to be
passed via the workflow. As such, the `docker` invocations themselves
will try to load the referenced library, which fails as it is only
installed within the container.

Fixes: #2786