Github: Fix bump-dependencies runs on existing PRs
The workflow got permission errors from the Github API when trying to
edit existing PRs which had already been added to an organization
project.
The reason for this is that the GITHUB_TOKEN of the run is scoped to the
repo, but `gh edit` tries to fetch all fields of a PR which includes the
inaccessible organization project field.
Therefore, use `gh api` instead which can be used in a more fine-grained
way.

Fixes: https://github.com/jamulussoftware/jamulus/pull/2777#issuecomment-1222793504