Build: Mac: Fix universal build to contain both archs
One of the last-minute changes broke the universal build and resulted in
a universal build which only contained a single architecture.
The reason was that the added `distclean` action removed the artifact
for the first architecture.

This change moves the intermediate artifacts to a temporary, already
defined directory which is definitely not touched by `make distclean`.

Related: https://github.com/jamulussoftware/jamulus/pull/2808#discussion_r956749658
Related: https://github.com/jamulussoftware/jamulus/pull/2808#issuecomment-1230842458