Util: Make CLocale::IsCountryCodeSupported more robust
On Qt6, we previously relied on the caller to only supply country codes within
bounds. With this change we properly check before risking an uninitialized array
access.

On Qt5, we did not do any checks at all. Now we ensure that we what is
provided is a valid Qt5 country code.

Related: #2809