tools/get_release_contributors: Add user lookup by email
This extends the release contributor scripts to look up users by email
via Github search/users API.
This is helpful for the Weblate process which only adds the original
authors as Co-authored-by.

In addition, the output changes to fall back to realnames (from the
Co-authored-by lines) if not Github login is found.
To make that mixed output more readable, commas are added between user
names now.